Move icons module to the main process
