+++ src/main/icons.spec.js
